+++ app/src/main/java/jp/co/charco/praisesreminder/PraiseInputBottomSheetFragment.kt
@@ -33,9 +33,11 @@ class PraiseInputBottomSheetFragment : BottomSheetDialogFragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        if (context is OnInputSubmitListener) {
-            listener = context as OnInputSubmitListener
-        } else throw IllegalStateException()
+        listener = when {
+            context is OnInputSubmitListener -> context as OnInputSubmitListener
+            parentFragment is OnInputSubmitListener -> parentFragment as OnInputSubmitListener
+            else -> throw IllegalStateException()
+        }
         // Keyboard で BottomSheet 隠れる問題の回避策 https://stackoverflow.com/a/50948146
         setStyle(DialogFragment.STYLE_NORMAL, R.style.BottomSheetDialog)
     }
